feat: ДАННЫЕ УДАЛЕНЫ

+++ js/clearData.js
@@ -1,4 +1,10 @@
 function clearData() {
-    localStorage.clear();
-    location.reload();
+    let forms = JSON.parse(localStorage.getItem('forms') || '[]');
+
+    if (forms.length === 0) {
+        iziToast.warning({title: '■■■■■■■■■', message: 'ДАННЫЕ УДАЛЕНЫ'});
+    } else {
+        localStorage.clear();
+        location.reload();
+    }
 }
